fix for cast on 64 bits platform
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@1942 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/SizeProperty.cpp
@@ -47,7 +47,7 @@ void SizeProperty::scale( const tlp::Vector<float,3>& v, Graph *sg ) {
 //=============================================================================
 Size SizeProperty::getMax(Graph *sg) {
   if (sg==0) sg=graph;
-  unsigned sgi=(unsigned long)sg;
+  unsigned long sgi=(unsigned long)sg;
   if (minMaxOk.find(sgi)==minMaxOk.end()) minMaxOk[sgi]=false;
   if (!minMaxOk[sgi]) computeMinMax(sg);
   return max[sgi];
